pyproject.toml
[project]
name = "hackathon-judge"
version = "0.1.0"
description = "Automated hackathon evaluation system"
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1.40",
    "requests>=2.31.0",
    "pandas>=2.1.0",
    "rich>=13.7.0",
    "click>=8.1.7",
]

[project.scripts]
hackathon-judge = "hackathon_judge.cli.main:cli"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
